# src/main/java/com/JobConnect/JobConnect/Model/Application.java
package com.JobConnect.JobConnect.Model;

import java.time.LocalDateTime;
import java.util.UUID;

import org.hibernate.annotations.GenericGenerator;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.persistence.Entity;
import jakarta.persistence.FetchType;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.Id;
import jakarta.persistence.JoinColumn;
import jakarta.persistence.ManyToOne;
import jakarta.persistence.Table;

@Entity
@Table(name = "applications")
@Schema(description = "Application entity representing a job application")
public class Application {
    @Id
    @GeneratedValue(generator = "UUID")
    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
    @Schema(description = "Unique identifier of the application", example = "123e4567-e89b-12d3-a456-426614174000")
    private UUID id;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "job_id")
    @JsonIgnoreProperties({"applications", "company"})
    @Schema(description = "Job being applied for")
    private Job job;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "applicant_id")
    @JsonIgnoreProperties({"applications", "experiences", "skills"})
    @Schema(description = "Applicant applying for the job")
    private Applicant applicant;

    @Schema(description = "Date and time when the application was submitted")
    private LocalDateTime appliedAt = LocalDateTime.now();

    // Default constructor required by JPA
    public Application() {}

    // Constructor with job and applicant
    public Application(Job job, Applicant applicant) {
        this.job = job;
        this.applicant = applicant;
    }

    // Getters and setters
    public UUID getId() { return id; }

    public Job getJob() { return job; }
    public void setJob(Job job) { this.job = job; }

    public Applicant getApplicant() { return applicant; }
    public void setApplicant(Applicant applicant) { this.applicant = applicant; }

    public LocalDateTime getAppliedAt() { return appliedAt; }
    public void setAppliedAt(LocalDateTime appliedAt) { this.appliedAt = appliedAt; }
}