# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: JobConnect API
  description: Job portal API for managing companies, jobs, applicants and applications
  version: 1.0.0
  contact:
    name: Omar Soliman
    email: omar.soliman@dauphine.eu
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Development Server
...
paths:
...
  /api/applicants:
    get:
      tags:
        - Applicants
      summary: Get all applicants
      responses:
        '200':
          description: List of applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Applicant'
    post:
      tags:
        - Applicants
      summary: Create applicant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
      responses:
        '201':
          description: Applicant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Applicant'
  /api/applicants/{applicantId}:
    parameters:
      - name: applicantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Applicants
      summary: Get applicant by ID
      responses:
        '200':
          description: Applicant found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Applicant'
        '404':
          description: Applicant not found
    put:
      tags:
        - Applicants
      summary: Update applicant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Applicant'
      responses:
        '200':
          description: Applicant updated
        '404':
          description: Applicant not found
    delete:
      tags:
        - Applicants
      summary: Delete applicant
      responses:
        '204':
          description: Applicant deleted
        '404':
          description: Applicant not found
  /api/applicants/{applicantId}/applications:
    parameters:
      - name: applicantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Applicants
        - Applications
      summary: Get applications by applicant
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
  /api/applicants/{applicantId}/experiences:
    parameters:
      - name: applicantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Applicants
        - Experiences
      summary: Get experiences by applicant
      responses:
        '200':
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Experience'
  /api/applications:
    get:
      tags:
        - Applications
      summary: Get all applications
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
    post:
      tags:
        - Applications
      summary: Create application
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
  /api/applications/{applicationId}:
    parameters:
      - name: applicationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Applications
      summary: Get application by ID
      responses:
        '200':
          description: Application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: Application not found
    delete:
      tags:
        - Applications
      summary: Delete application
      responses:
        '204':
          description: Application deleted
        '404':
          description: Application not found
components:
  schemas:
...
    Applicant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        experiences:
          type: array
          items:
            $ref: '#/components/schemas/Experience'
      required:
        - firstName
        - lastName
        - email
    ApplicantRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    Application:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicant:
          $ref: '#/components/schemas/ApplicantRef'
        job:
          $ref: '#/components/schemas/JobRef'
        appliedAt:
          type: string
          format: date-time
      required:
        - applicant
        - job
...
    JobRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required:
        - id
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - name
    Experience:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        role:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        applicant:
          $ref: '#/components/schemas/ApplicantRef'
      required:
        - companyName
        - role
        - startDate 